Home

MapImages.jl

MapImages.jl

Index

Index

MapImages.coarse_grid

MapImages.crop_rsc_data

MapImages.find_obj

MapImages.get_cache_dir

MapImages.grid

MapImages.grid_extent

MapImages.intersection_corners

MapImages.latlon_to_dist

MapImages.nsew

MapImages.read_station_llas

MapImages.rowcol_to_latlon

MapImages.station_lonlat

MapImages.coarse_grid — Function.

Create a grid from a DemRsc on the 10^(-digits) grid lines

source

MapImages.crop_rsc_data — Function.

Adjusts the old demrsc of a MapImage after slicing

Takes the 'rows' and 'cols' from demrsc and adjusts for the smaller size with a new dict

The row/colstart are the indices where we have sliced: e.g. img[2:3:11, :] would have rowstart = 2, rowstep = 3, newrows = 4

source

MapImages.find_obj — Method.

A = find_obj(As) returns the first MapImage among the arguments.

source

MapImages.get_cache_dir — Function.

Returns the config folder for the application. The default behavior is to return whatever is most appropriate for the operating system.

This is used to store gps timeseries data

the following folders could be returned: Mac OS X: $~/Library/Application Support/apertools$ Mac OS X (POSIX): $~/.apertools$ Unix: $~/.cache/apertools$ Unix (POSIX): $~/.apertools$

force_posix: if this is set to true then on any POSIX system the folder will be stored in the home folder with a leading dot instead of the XDG config home or darwin's application support folder.

Base on: https://github.com/pallets/click/blob/master/click/utils.py#L368

source

MapImages.grid — Method.

Returns meshgrid-like arrays X, Y

julia> XX, YY = MapImages.grid(demrsc);

julia> XX[1:3, 1:3] 3×3 Array{Float64,2}: -104.0 -103.998 -103.997 -104.0 -103.998 -103.997 -104.0 -103.998 -103.997

julia> YY[1:3, 1:3] 3×3 Array{Float64,2}: 31.9 31.9 31.9 31.8983 31.8983 31.8983 31.8967 31.8967 31.8967

julia> XX, YY = MapImages.grid(demrsc, sparse=true);

source

MapImages.grid_extent — Method.

Get the boundaries of a grid

Returns: tuple[float]: the boundaries of the latlon grid in order: (lonleft,lonright,latbottom,lattop)

source

MapImages.intersection_corners — Method.

Returns: tuple[float]: the boundaries of the intersection box of the 2 areas in order: (lonleft,lonright,latbottom,lattop)

source

MapImages.latlon_to_dist — Function.

Find the distance between two lat/lon points on Earth

latlon_to_dist(lat_lon_start, lat_lon_end, R=6378)

Uses the haversine formula: https://en.wikipedia.org/wiki/Haversine_formula so it does not account for the ellopsoidal Earth shape. Will be with about 0.5-1% of the correct value.

Notes: lats and lons are in degrees, and the values used for R Earth (6373 km) are optimized for locations around 39 degrees from the equator

Reference: https://andrew.hedges.name/experiments/haversine/

source

MapImages.nsew — Method.

return tuple of (north, south, east, west) from demrsc data

source

MapImages.read_station_llas — Method.

Read in a DataFrame of gps stations with columns name, lat, lon, alt

source

MapImages.rowcol_to_latlon — Method.

Takes the row, col of a pixel and finds its lat/lon

source

MapImages.station_lonlat — Method.

Return the (lon, lat) of a station_name

source