Home

MapImages.jl

MapImages.jl

Index

Index

MapImages.crop_demrsc

MapImages.find_obj

MapImages.get_cache_dir

MapImages.grid

MapImages.grid_extent

MapImages.intersection_corners

MapImages.km_to_deg

MapImages.latlon_to_dist

MapImages.nsew

MapImages.read_station_llas

MapImages.rowcol_to_latlon

MapImages.station_lonlat

MapImages.@allow_mapimage

MapImages.crop_demrsc — Method.

crop_demrsc(demrsc::DemRsc;
            new_rows=demrsc.rows, new_cols=demrsc.cols, 
            row_start=1, col_start=1, 
            row_step=1, col_step=1)

Adjusts the old demrsc of a MapImage after slicing. Takes the 'rows' and 'cols' from demrsc and adjusts for the smaller size with a new dict

The row_start/col_start are the indices where we have sliced. This means:

img[2:3:11, :]

would have row_start = 2, row_step = 3, new_rows = 4

source

MapImages.find_obj — Method.

A = find_obj(As) returns the first MapImage among the arguments.

source

MapImages.get_cache_dir — Function.

Returns the config folder for the application. The default behavior is to return whatever is most appropriate for the operating system.

This is used to store gps timeseries data

the following folders could be returned: Mac OS X: $~/Library/Application Support/apertools$ Mac OS X (POSIX): $~/.apertools$ Unix: $~/.cache/apertools$ Unix (POSIX): $~/.apertools$

force_posix: if this is set to true then on any POSIX system the folder will be stored in the home folder with a leading dot instead of the XDG config home or darwin's application support folder.

Base on: https://github.com/pallets/click/blob/master/click/utils.py#L368

source

MapImages.grid — Method.

grid(demrsc::DemRsc; sparse=false)

Returns meshgrid-like arrays X, Y

Examples

julia> demrsc = DemRsc(x_first=1.0, y_first=2.0, x_step=0.1, y_step=-0.2, file_length=6, width=5);

julia> XX, YY = MapImages.grid(demrsc);

julia> XX[1:3, 1:3]
3×3 Array{Float64,2}:
 1.0  1.1  1.2
 1.0  1.1  1.2
 1.0  1.1  1.2
 
julia> YY[1:3, 1:3]
3×3 Array{Float64,2}:
 2.0  2.0  2.0
 1.8  1.8  1.8
 1.6  1.6  1.6

julia> size(XX), size(YY)
((6, 5), (6, 5))

julia> XX, YY = MapImages.grid(demrsc, sparse=true);

julia> size(XX), size(YY)
((5,), (6,))

source

MapImages.grid_extent — Method.

grid_extent(demrsc::DemRsc)

Get the boundaries of a grid

Returns: tuple[float]: the boundaries of the latlon grid in order: (lonleft,lonright,latbottom,lattop)

source

MapImages.intersection_corners — Method.

Returns: tuple[float]: the boundaries of the intersection box of the 2 areas in order: (lonleft,lonright,latbottom,lattop)

source

MapImages.km_to_deg — Function.

Find the degrees separation for distance km assuming distance along great circle arc

source

MapImages.latlon_to_dist — Function.

Find the distance between two lat/lon points on Earth

latlon_to_dist(lat_lon_start, lat_lon_end, R=6378)

Uses the haversine formula: https://en.wikipedia.org/wiki/Haversine_formula so it does not account for the ellopsoidal Earth shape. Will be with about 0.5-1% of the correct value.

Notes: lats and lons are in degrees, and the values used for R Earth (6373 km) are optimized for locations around 39 degrees from the equator

Reference: https://andrew.hedges.name/experiments/haversine/

source

MapImages.nsew — Method.

return tuple of (north, south, east, west) from demrsc data

source

MapImages.read_station_llas — Method.

Read in a DataFrame of gps stations with columns name, lat, lon, alt

source

MapImages.rowcol_to_latlon — Method.

rowcol_to_latlon(demrsc::DemRsc, row, col)

Takes the row, col of a pixel and finds its lat/lon within the grid

source

MapImages.station_lonlat — Method.

Return the (lon, lat) of a station_name

source

MapImages.@allow_mapimage — Macro.

@allow_mapimage(funcname)

macro to make DemRsc functions equivalently defined for MapImage.

The macro just passes along the MapImage.demrsc to the function already defined

Usage

nearest_row(demrsc::DemRsc, lat::AbstractFloat) = Int(round(1 + (lat - d.y_first) / d.y_step))
@allow_mapimage nearest_row

This is the same as:

nearest_row(img::MapImage, lat) = nearest_row(img.demrsc, lat)

source